requirements.txt
pigpio
flask
adafruit-circuitpython-ssd1306
adafruit-circuitpython-dht
pillow
adafruit-blinka